Onnx refactored

+++ pt2keras/core/onnx/graph.py
@@ -2,6 +2,7 @@ import logging
 import os.path
 import typing as t
 import onnx
+import onnxruntime as ort
 import torch.nn as nn
 
 from collections import OrderedDict
@@ -10,14 +11,12 @@ import torch.onnx
 from onnx.helper import printable_node
 
 from onnx import numpy_helper
-from onnx.helper import get_attribute_value
 from tensorflow import keras
 
-from .util import get_graph_input_shape, test_model_output
+from .util import get_graph_input_shape, get_graph_output_shape, test_model_output
 
 
 class Graph:
-
     _SUPPORTED_OPERATIONS = {}
     _LOGGER = logging.getLogger('onnx::Graph')
 
@@ -25,12 +24,11 @@ class Graph:
     A class that encapsulates and converts the onnx graph representation into
     Keras format
     """
+
     def __init__(self,
-                 pytorch_model: nn.Module,
+                 model: t.Union[nn.Module, str, onnx.ModelProto],
                  input_shape: t.Tuple,
-                 source_format: t.Tuple = ('B', 'C', 'H', 'W'),
-                 target_format: t.Tuple = ('B', 'H', 'W', 'C'),
-                 opset_version: int = 13):
+                 opset_version: int = 12):
         """
         By default the onnx graph is designed to convert PyTorch onnx
         models to Keras. By making small modifications and writing converters,
@@ -39,49 +37,65 @@ class Graph:
         For now, since I am working with computer vision models, this onnx converter
         is designed to work with models in the computer vision fields, that generally
         have at most, 4 dimensional inputs
-
-        Args:
-            source_format: The source input format
-            target_format:
         """
-        self.pytorch_model = pytorch_model
         self.pytorch_input_shape = input_shape
-        self.source_format = source_format
-        self.target_format = target_format
         self.opset_version = opset_version
 
-        dummy_input = torch.randn(input_shape)
-        output = self.pytorch_model(dummy_input)
-        output_names = []
-        if isinstance(output, (t.Tuple, t.List)):
-            for i in range(len(output)):
-                output_names.append(f'output_{i}')
+        # Added as class object for unit testing
+        self.test_stats = None
+        self.model = model
+
+        if isinstance(self.model, nn.Module):
+            # For now, we assume that there is only a single input
+            # We can later change this to support multiple inputs
+            dummy_input = torch.randn(input_shape)
+            output = self.model(dummy_input)
+            self.output_names = []
+            if isinstance(output, (t.Tuple, t.List)):
+                for i in range(len(output)):
+                    self.output_names.append(f'output_{i}')
+            else:
+                self.output_names.append('output_0')
+
+            hash_str = f'__{hash(model)}__.onnx'
+            torch.onnx.export(self.model,
+                              dummy_input,
+                              hash_str,
+                              do_constant_folding=True,  # whether to execute constant folding for optimization
+                              verbose=False,
+                              opset_version=self.opset_version,
+                              export_params=True,
+                              input_names=['input_0'],
+                              output_names=self.output_names)
+
+            # Check model and create onnx runtime session
+            self.onnx_model = onnx.load(hash_str)
+            self.ort_session = ort.InferenceSession(hash_str)
+            onnx.checker.check_model(self.onnx_model)
+            if os.path.exists(hash_str):
+                os.remove(hash_str)
+
+            # Test model for equality
+            self._LOGGER.debug('Testing for equality --- PyTorch vs Onnx')
+            # output_ort = self.ort_session.run(None, {"input_0": dummy_input})
+            # output_pt = output.detach().cpu().numpy()
+        elif isinstance(self.model, str):
+            self.onnx_model = onnx.load(model)
+            onnx.checker.check_model(self.onnx_model)
+            self.model = ort.InferenceSession(self.onnx_model.SerializeToString())
         else:
-            output_names.append('output_0')
+            raise ValueError(f'Invalid model type: {self.model}. Please pass in '
+                             f'PyTorch model or onnx model string path.')
 
-        hash_str = f'__{hash(pytorch_model)}__.onnx'
-        torch.onnx.export(self.pytorch_model,
-                          dummy_input,
-                          hash_str,
-                          do_constant_folding=True,  # whether to execute constant folding for optimization
-                          verbose=False,
-                          opset_version=self.opset_version,
-                          # training=TrainingMode.TRAINING,
-                          export_params=True,
-                          input_names=['input_0'],
-                          output_names=output_names)
+        # For now, assume we are working with models that have a single input.
+        # we will later need to test this with models that have multiple inputs
+        self.output_names = [output['name'] for output in get_graph_output_shape(self.onnx_model.graph)]
 
-        # Check model
-        self.onnx_model = onnx.load(hash_str)
-        onnx.checker.check_model(self.onnx_model)
-        if os.path.exists(hash_str):
-            os.remove(hash_str)
-
-        self.weights = OrderedDict()
         self.node_dict = OrderedDict()
-
         # The computational graph value we are building up
         self.computational_graph = {}
+        # This is for accessing keras input cache
+        self.forward_input_cache = {}
 
         # initialization phase:
         # ------------------------------------------------------
@@ -122,18 +136,6 @@ class Graph:
             node_info = printable_node(node)
             self._LOGGER.debug(f'NODE::::: {node_info}')
 
-            if node.attribute:
-                for attribute in node.attribute:
-                    # INTS, FLOATS, etc
-                    type_data = get_attribute_value(attribute)
-                    # Try casting to numpy if possible
-                    try:
-                        data = numpy_helper.to_array(type_data)
-                        onnx_node_obj.attributes[attribute.name] = data
-                    # just save raw representation
-                    except:
-                        onnx_node_obj.attributes[attribute.name] = type_data
-
             # Constant nodes have no input and just single output
             # we need to extract constants
             if node.op_type == 'Constant':
@@ -148,25 +150,39 @@ class Graph:
         input_shape = self.pytorch_input_shape
 
         # Change image shape from BCHW to BHWC (TensorFlow / Keras default shape)
-        dims = (i for i in range(len(input_shape))) if len(input_shape) != 4 else (0, 2, 3, 1)
+        dims = (i for i in range(len(input_shape))) if len(input_shape) != 4 or input_shape[-1] == 3 else (0, 2, 3, 1)
 
         # For now, assume we are working with models that have a single input.
         # we will later need to test this with models that have multiple inputs
-        input_shape = get_graph_input_shape(self.onnx_model.graph, dims)[0]['shape']
+        input_data = get_graph_input_shape(self.onnx_model.graph, dims)[0]
+        input_shape = input_data['shape']
+        input_name = input_data['name']
+
+        # First dimension = 0 is dynamic batching. We disallow those ...
+        if input_shape[0] == 0:
+            error_msg = 'Cannot convert model with dynamic batch size ... '
+            self._LOGGER.error(error_msg)
+            raise ValueError(error_msg)
 
         # Create input object to feed to the model.
         # This will need to change in the future if we want to support multiple inputs
         inputs = keras.Input(batch_shape=input_shape)
+        self.forward_input_cache[input_name] = inputs
         outputs = inputs
 
         # Store unsupported operations in a set so that
         # we can figure out what we operations we need to add in the future
         has_unsupported_ops = False
         unsupported_ops = set()
+        output_list = []
+
+        # Create a new test stats object for each conversion run.
+        self.test_stats = TestResults()
 
         # Convert the model
         for node_key, node in self.node_dict.items():
             op_type = node.op_type
+
             if op_type not in self._SUPPORTED_OPERATIONS:
                 unsupported_ops.add(op_type)
                 has_unsupported_ops = True
@@ -175,77 +191,164 @@ class Graph:
             # does not have all the necessary converters.
             # Users can extend the library by adding the required support
             # without modifying the library directly
-            if has_unsupported_ops:
+            # -----------------------------------------------
+            # Constant nodes have no input
+            if has_unsupported_ops or node.name.startswith('Constant'):
                 continue
 
             conversion_func = self._SUPPORTED_OPERATIONS[op_type]
             node_inputs = []
+
+            # add inputs
             for input_node in node.input_nodes:
                 if input_node in self.computational_graph:
                     node_inputs.append(self.computational_graph[input_node])
 
+            input_node_name = node.input_nodes[0]
+            # retrieve previous layer
+            if input_node_name in self.forward_input_cache:
+                input_keras_layer = self.forward_input_cache[input_node_name]
+            # Is a weight node
+            else:
+                input_keras_layer = self.computational_graph[input_node_name]
+
             # Convert to keras
-            outputs = conversion_func(node, outputs, self.computational_graph, *node_inputs)
+            outputs = conversion_func(node, input_keras_layer, self.computational_graph,
+                                      self.node_dict, self.test_stats, *node_inputs)
+
+            self.forward_input_cache[node.name] = outputs
+
+            # Add to output data if output_node found
+            if node.output_nodes[0] in self.output_names:
+                output_list.append(outputs)
+
             Graph._LOGGER.info(f'Successfully converted: {node}')
-            Graph._LOGGER.info(f'Outputs: {node}')
+
+        # Print model conversion result if needed
+        self._LOGGER.info('Model conversion report: ###################### \n'
+                          f'{self.test_stats.get_test_results()}')
 
         if has_unsupported_ops:
             unsupported_operations = "\n- ".join(unsupported_ops)
             raise ValueError('Failed to convert model. The following operations are currently unsupported: '
                              f'{unsupported_operations}')
 
+        if len(output_list) > 1:
+            outputs = output_list
+
         model = keras.Model(inputs, outputs)
         # Test the Keras model output.
         # Error will be asserted if the output dimensions or values are very different.
-        test_model_output(self.pytorch_model, model, self.pytorch_input_shape, input_shape)
+        test_model_output(self.model, model, self.pytorch_input_shape, input_shape)
         return model
 
     def _initialize_weights(self):
         for weight in self.onnx_model.graph.initializer:
             name = weight.name
             np_weights = numpy_helper.to_array(weight)
-            if len(np_weights.shape) == len(self.source_format):
-                Graph._LOGGER.info(f'Transposing weights: {name}')
-                # H,W,IC,OC
-                np_weights = np_weights.transpose([2, 3, 1, 0])
-
-            # Note tht we can also check whether the initialize is actually a weight
-            # or a constant by checking what the name endswith
-            # is_weight = name.endswith('weight') or name.endswith('bias')
 
             # operations such as the division in
             # (output + 6 * 3) / 3
             # can be considered an initializer
             # in this case, we need to add a constant node to the graph
             # if not is_weight:
-            self.computational_graph[name] = np_weights
-            self.weights[name] = {
-                'weights': np_weights
-            }
 
-        # move data to nodes
+            # IMPORTANT: Note that computational graph also contains
+            # not only weights, but constant values from constant nodes such as
+            # when we do element-wise additional to a constant
+            self.computational_graph[name] = np_weights
+
+        # move Weights to node for convenience
         for node in self.node_dict.values():
             for input_node in node.input_nodes:
-                if input_node in self.weights:
-                    node.weights.append(self.weights[input_node]['weights'])
-
-        # Remove temporary dat
-        del self.weights
+                if input_node in self.computational_graph:
+                    node.weights.append(self.computational_graph[input_node])
 
         self._LOGGER.info(f'Built computational graph: {self.computational_graph.keys()}')
+
+
+class TestResults:
+    """
+    The class contains test results when converting the model on a node-by-node basis
+    """
+
+    def __init__(self):
+        self.tested_nodes: t.List = []
+        self.untested_nodes: t.List = []
+
+        self.tested_operations: t.Set = set()
+        self.untested_operations: t.Set = set()
+
+    def get_test_results(self, stat_delimiter: str = '\n - ') -> str:
+        """
+        Get a formatted string visualizing the test results.
+        Args:
+            stat_delimiter: The delimiter for each entry.
+            The default format is '-' delimited bullet points.
+            E.g.
+            - point1
+            - point2
+
+        Returns:
+            The formatted string containing the test results representing the
+            conversion results of a target model.
+        """
+        # Node-related stats
+        node_count = len(self.tested_nodes) + len(self.untested_nodes)
+        tested_nodes = stat_delimiter + stat_delimiter.join(self.tested_nodes)
+        untested_nodes = stat_delimiter + stat_delimiter.join(self.untested_nodes)
+
+        # Operation states
+        tested_operations = stat_delimiter + stat_delimiter.join(self.tested_operations)
+        untested_operations = stat_delimiter + stat_delimiter.join(self.untested_operations)
+
+        return f'Total node count: {node_count}. \n' \
+               f'Test nodes: {tested_nodes} \n' \
+               f'Untested nodes: {untested_nodes}. \n' \
+               f'Tested operations: {tested_operations} \n' \
+               f'Untested operations: {untested_operations} \n'
+
+
+def onnx_node_attributes_to_dict(attributes):
+    """
+    From: https://github.com/gmalivenko/onnx2keras/blob/master/onnx2keras/converter.py
+    Parse ONNX attributes to Python dictionary
+    Args:
+        attributes: The attributes from OnnxNode
+    """
+    def onnx_attribute_to_dict(onnx_attr):
+        """
+        Parse ONNX attribute
+        Args:
+            onnx_attr: ONNX attribute
+        Returns:
+            Python data type
+        """
+        if onnx_attr.HasField('t'):
+            return numpy_helper.to_array(getattr(onnx_attr, 't'))
+
+        for attr_type in ['f', 'i', 's']:
+            if onnx_attr.HasField(attr_type):
+                return getattr(onnx_attr, attr_type)
+
+        for attr_type in ['floats', 'ints', 'strings']:
+            if getattr(onnx_attr, attr_type):
+                return list(getattr(onnx_attr, attr_type))
+    return {arg.name: onnx_attribute_to_dict(arg) for arg in attributes}
 
 
 class OnnxNode:
     """
     A node in the onnx graph
     """
+
     def __init__(self, node):
         self.name = node.name
         self.op_type = node.op_type
         self.input_nodes = node.input
         self.output_nodes = node.output
         self.weights = []
-        self.attributes = {}
+        self.attributes = onnx_node_attributes_to_dict(node.attribute)
 
     def __repr__(self):
         return f'name: {self.name}: ' \
